#77 create test (#78)
* #77 indicatorとquestionListのunitテスト作成

* #77 answerList_unitテストの作成

* #77 snowboardテストの作成

* #77 RideTypeとResultのテストの作成

* #77 pageStateとThemeStatusのテストの作成

* #77 resultDbテストの作成

* #77 診断結果表示widgetテストの作成

* #77 インテグレーションテストの作成

+++ lib/model/answer.dart
@@ -6,6 +6,6 @@ part 'answer.freezed.dart';
 class Answer with _$Answer {
   const factory Answer({
     required String category,
-    required double answerValue,
+    required int answerValue,
   }) = _Answer;
 }
